+++ examples/index.android.js
@@ -1,0 +1,10 @@
+/**
+ * Sample Elm Native UI app
+ * https://github.com/elm-native-ui/elm-native-ui
+ * https://github.com/facebook/react-native
+ */
+'use strict';
+var {AppRegistry} = require('react-native');
+var ElmAppWrapper = require('./ElmAppWrapper');
+
+AppRegistry.registerComponent('ElmNative', ElmAppWrapper);
