--- index.android.js
@@ -1,9 +1,0 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- */
-'use strict';
-var {AppRegistry} = require('react-native');
-var ElmAppWrapper = require('./ElmAppWrapper');
-
-AppRegistry.registerComponent('ElmNative', ElmAppWrapper);
